requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.23.2
python-multipart==0.0.9
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
transformers==4.39.3
torch==2.1.2
easyocr==1.7.1
jinja2==3.1.3
sentencepiece==0.1.99
PyMuPDF>=1.24.0
img2pdf>=0.5.1
manga-ocr>=0.1.8
